% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_tree_ransac.R
\name{individual_tree_ransac}
\alias{individual_tree_ransac}
\title{Function updates a tree ransac fit in a given data frame}
\usage{
individual_tree_ransac(
  data,
  tree_id,
  las_slice,
  iterations,
  threshold,
  inclusion
)
}
\arguments{
\item{data}{a data frame object that contains tree ransac fits from las_slice_circle_fitting function}

\item{tree_id}{the tree id to preform ransac circle fitting}

\item{las_slice}{A slice from dbscan that has points labled with tree ids}

\item{iterations}{the max number of iterations to preform ransac circle fits}

\item{threshold}{the minimum distance a point needs to be within the circle circumference to be considered an inlier}

\item{inclusion}{the desired inclusion to be considered a good circle fit to the data}
}
\description{
Takes in a data frame of tree fits from las_slice_circle_fitting, a tree id and ransac parameters.
Does a ransac fit for the selected tree and updates the returned dataframe
}
\details{
See las_slice_circle_fitting for information on how the fit is performed
}
