% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/las_slice_circle_fitting.R
\name{las_slice_circle_fitting}
\alias{las_slice_circle_fitting}
\title{slice circle fit (PCA / bisecting cords)}
\usage{
las_slice_circle_fitting(las_slice, iterations, threshold, inclusion)
}
\arguments{
\item{las_slice}{A slice from dbscan that has points labled with tree ids}

\item{iterations}{the max number of iterations to preform ransac circle fits}

\item{threshold}{the minimum distance a point needs to be within the circle circumference to be considered an inlier}

\item{inclusion}{the desired inclusion to be considered a good circle fit to the data}
}
\description{
this function takes in a las slice from dbscan that has labled points to tree ids and returns a dataframe
of tree ids, the x, y, and z corrdiantes for the circle fit on a slice, the radius fo the circle fit, the
mean squared error of the points and the percent inclusion of each point within the tree
}
\details{
It preforms ransac circle fitting by:
\enumerate{
\item transform the las objest into a matrix
\item loop through the tree ids one at a time
\item prefrom robust pca using package rospca to find tree slice centers and eigenvectors(loadings)
\item use center and loadings to project the 3d point cloud onto a 2d principal component space
\item call the rcpp function ransac circle fit to preform circle fitting in parallel with given parameters
\item return a data frame that contains tree information
}
}
